+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@
 - add support for generalised expression trees to the algorithm configuration model (conditionals, forbidden combinations, general assertions)
 - add support for automatically computed parameter values
 - add live command display during execution
+- add support for Gradle test suites
 
 **Changed**